requirements.txt
gensim==4.2.0
GitPython==3.1.30
hdbscan==0.8.29
humanize==4.4.0
matplotlib==3.6.3
nltk==3.7
numpy==1.23.5
pandas==1.5.1
psutil==5.9.4
PyGithub==1.57
pynndescent==0.5.8
s_dbw==0.4.0
scikit_learn==1.2.1
scipy==1.10.0
sentence_transformers==2.2.2
sklearn_som==1.1.0
spacy==3.4.3
sparse_som==0.6.1
statsmodels==0.13.5
tqdm==4.64.1
umap_learn==0.5.3
